Allowing for Incremental Tweens
Allows the developer to call TWEEN.update manually.  When initializing
a new Tween object, add the FPS as the 2nd parameter, which will force
the Tween library to increment the tween by one single tick (instead
of figuring out how much it should have changed according to the duration
since the last update).